+++ app/Models/BindingQueue.php
@@ -25,6 +25,11 @@ class BindingQueue extends Model
         return $this->belongsTo(OrderedBook::class, 'ordered_book_id', 'ordered_book_id');
     }
 
+    public function coverPrintingQueue()
+    {
+        return $this->belongsTo(coverPrintingQueue::class, ['order_id', 'ordered_book_id'], ['order_id', 'ordered_book_id']);
+    }
+
     public function handler()
     {
         return $this->belongsTo(Employee::class, 'handled_by');