+++ database/migrations/2025_05_30_140034_create_qc_queues_table.php
@@ -20,6 +20,7 @@ class CreateQcQueuesTable extends Migration
 
             $table->index('order_id');
             $table->index('ordered_book_id');
+            $table->foreign('order_id')->references('order_id')->on('orders')->onDelete('cascade');
             $table->foreign('handled_by')->references('id')->on('employees')->onDelete('restrict');
         });
     }