+++ changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.0.3]
+### Added
+- Add View Features For Order
+- Add Delet Features For Order
+- View Cover Type in Design Queue
+- Add Lamination Type Property for Ordered Books
+
+### Changed
+- Show Cover Print status in Binding Queue in stead of Cover Print/Print queue
+
+
 ## [1.0.2] - 2025-06-10
 ### Added
 - Add Updating Features For Order